improve articles reactions error handeling

+++ articles_app/api/v1/views.py
@@ -45,48 +45,54 @@ class ArticleReactionsAPIView(APIView):
             reaction = serializer.validated_data.get('reaction')
             user_ip = get_ip(request)
 
+            # Checks if article exists
+            if ArticleModel.objects.filter(id=article_id).exists():
             # If the user clicks on the like button
-            if reaction == 'like':
+                if reaction == 'like':
 
-                # Checks if the user has already liked the article, does nothing (You can customize it to delete the previous like if the user had already liked the article)
-                if ArticleLikesModel.objects.filter(user=user_ip, article_id=article_id).exists():
-                    response = {
-                        'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
-                        'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count()
-                    }
-
-                else:
-                    # If the user had already disliked the article, it deletes it
-                    if ArticleDisLikesModel.objects.filter(user=user_ip, article_id=article_id).exists():
-                        ArticleDisLikesModel.objects.filter(user=user_ip, article_id=article_id).delete()
-
-                    # Record a new like for article
-                    ArticleLikesModel.objects.create(user=user_ip, article_id=article_id)
-                    response = {
-                        'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
-                        'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count()
-                   }
-            # If the user clicks on the dislike button
-            elif reaction == 'dislike':
-                # Checks if the user has already disliked the article, does nothing (You can customize it to delete the previous dislike if the user had already disliked the article)
-                if ArticleDisLikesModel.objects.filter(user=user_ip, article_id=article_id).exists():
-                    response = {
-                        'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
-                        'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count()
-                    }
-                else:
-                    # If the user had already liked the article, it deletes
+                    # Checks if the user has already liked the article, does nothing (You can customize it to delete the previous like if the user had already liked the article)
                     if ArticleLikesModel.objects.filter(user=user_ip, article_id=article_id).exists():
-                        ArticleLikesModel.objects.filter(user=user_ip, article_id=article_id).delete()
+                        response = {
+                            'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
+                            'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count(),
+                        }
 
-                    # Record a new dislike for article
-                    ArticleDisLikesModel.objects.create(user=user_ip, article_id=article_id)
-                    response = {
-                        'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
-                        'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count()
-                    }
+                    else:
+                        # If the user had already disliked the article, it deletes it
+                        if ArticleDisLikesModel.objects.filter(user=user_ip, article_id=article_id).exists():
+                            ArticleDisLikesModel.objects.filter(user=user_ip, article_id=article_id).delete()
 
-            return Response(response, status=status.HTTP_200_OK)
+                        # Record a new like for article
+                        ArticleLikesModel.objects.create(user=user_ip, article_id=article_id)
+                        response = {
+                            'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
+                            'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count()
+                       }
+                # If the user clicks on the dislike button
+                elif reaction == 'dislike':
+                    # Checks if the user has already disliked the article, does nothing (You can customize it to delete the previous dislike if the user had already disliked the article)
+                    if ArticleDisLikesModel.objects.filter(user=user_ip, article_id=article_id).exists():
+                        response = {
+                            'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
+                            'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count()
+                        }
+                    else:
+                        # If the user had already liked the article, it deletes
+                        if ArticleLikesModel.objects.filter(user=user_ip, article_id=article_id).exists():
+                            ArticleLikesModel.objects.filter(user=user_ip, article_id=article_id).delete()
+
+                        # Record a new dislike for article
+                        ArticleDisLikesModel.objects.create(user=user_ip, article_id=article_id)
+                        response = {
+                            'like': ArticleLikesModel.objects.filter(article_id=article_id).count(),
+                            'dislike': ArticleDisLikesModel.objects.filter(article_id=article_id).count()
+                        }
+
+                return Response(response, status=status.HTTP_200_OK)
+
+            # Checks if article not exists return 404
+            else:
+                return Response(serializer.errors, status=status.HTTP_404_NOT_FOUND)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
